fix(viewer): pin the portal guard, clamp the flip-up, extract SearchableSelect (#1924)
Review feedback from @louistrue on #1958.

**The guard that prevents the canonical portal regression was not pinned.**
`popupRef.current?.contains(target)` stops an inside-popup mousedown from
closing the popup and unmounting the option before its `click` fires — but
the commit test dispatched `click` with no preceding `mousedown`, so the
breaking sequence never ran and deleting the guard kept the suite green.
The test now dispatches `mousedown` then `click`, each in its own `act()`:
wrapping both in a single `act()` was itself part of the problem, since React
batched the mousedown's state update and deferred the unmount past the click,
hiding the bug even with the right sequence. Removing the guard now fails it.

Pointed out with the note that the previous commit's message was about
repairing a test that tested nothing — same shape, one file over.

**Flip-up could push the filter input off-screen irrecoverably.** The anchor
flipped when `spaceAbove > spaceBelow` without checking that `spaceAbove`
could actually hold the popup, and nothing clamped `max-h-[200px]` to the
space available. In a 300px trigger window with the trigger at top 160 /
bottom 190 the popup extended from y=158 to y=-42, putting the filter input
above the viewport with `position: fixed` making it unrecoverable — and a
300px window is exactly the popped-out panel this PR set out to support.

`computeSearchableSelectAnchor` now returns a `maxHeight` clamped to the
space actually available less an 8px edge margin, applied as an inline style
so it beats the Tailwind class. Chose the clamp over requiring
`spaceAbove >= popupMaxHeight` because it also fixes the not-flipped case
where space below is tight. A test using the exact scenario above asserts the
popup's top edge stays on-screen, and fails with `got -42` when reverted.

**Extracted `SearchableSelect` into its own module.** It and its helpers were
exported from `LensPanel.tsx` purely so tests could reach them, which forced
those tests to import the whole panel graph including a Zustand store that
touches `localStorage` at module init. The import-order warning that
documented that fragility is deleted, because the fragility is gone: the
anchor test file now runs standalone (9/9). `LensPanel.tsx` drops 1801 → 1550
lines; the new module is 303.

The shared portal hook unifying this with `combo-input.tsx`, the
`pointerEvents` divergence, and `RuleEditor`'s stacking context remain out of
scope as agreed — though the extraction does make that hook easier, since the
pure anchor/window/document helpers are now store-independent.

Verification: viewer suite 2177 tests / 2175 passed / 0 failed / 2 skipped,
`pnpm typecheck` 33/33, `check-test-wiring` clean. Both fixes mutation-tested
independently: removing the guard fails one test, reverting the clamp fails
two.
