+++ apps/viewer/src/components/viewer/LensPanel.tsx
@@ -17,15 +17,14 @@
  * classification systems, materials). No hardcoded IFC class lists.
  */
 
-import { memo, useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
-import { X, EyeOff, Palette, Check, Plus, Trash2, Pencil, Copy, Save, Download, Upload, Sparkles, Search, ChevronDown, ArrowUpDown, GripVertical } from 'lucide-react';
+import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { X, EyeOff, Palette, Check, Plus, Trash2, Pencil, Copy, Save, Download, Upload, Sparkles, ArrowUpDown, GripVertical } from 'lucide-react';
 import { discoverDataSources } from '@ifc-lite/lens';
+import { SearchableSelect } from './SearchableSelect';
 import { Button } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
 import { downloadFile } from '@/lib/export/download';
 import { tourAnchor, TOUR_ANCHORS, lensCardAnchor } from '@/lib/tours/anchors';
-import { usePortalContainer } from '@/components/ui/portal-container';
 import { useViewerStore } from '@/store';
 import { useLens } from '@/hooks/useLens';
 import { createLensDataProvider } from '@/lib/lens';
@@ -59,256 +58,6 @@ const TYPE_LABELS: Record<string, string> = {
 
 interface LensPanelProps {
   onClose?: () => void;
-}
-
-// ─── Searchable dropdown (for large dynamic lists) ──────────────────────────
-
-/** Popup's own CSS `max-h`, kept in one place so the flip math agrees with the class below. */
-const SEARCHABLE_SELECT_POPUP_MAX_HEIGHT = 200;
-
-/** Fixed-position coordinates for the portaled popup, computed from the trigger's rect. */
-interface SearchableSelectAnchor {
-  left: number;
-  width: number;
-  /** Distance from the viewport top to the trigger's bottom (used when opening down). */
-  top: number;
-  /** Distance from the viewport bottom to the trigger's top (used when opening up). */
-  bottom: number;
-  openUp: boolean;
-}
-
-/**
- * Resolve the `Window` that owns `el`, for viewport-relative measurements
- * and scroll/resize listeners. A panel popped out into its own OS / PiP
- * window (#1208) renders `SearchableSelect`'s trigger in *that* window's
- * document, not the main tab's — using the global `window` there computes
- * the flip decision against the wrong viewport height and attaches
- * scroll/resize listeners to the wrong window, so the popup never
- * repositions when the child window scrolls or resizes.
- *
- * `ownerDocument.defaultView` is the standard way to recover an element's
- * window; it is `null` only for a document detached from any window (e.g.
- * one created via `DOMParser` or `document.implementation.createDocument`,
- * never attached to a `<iframe>`/window). A mounted, visible trigger can't
- * be in a detached document, but we fall back to the global `window` rather
- * than asserting, since a stale ref during teardown is a more plausible way
- * to hit this than a truly detached document.
- */
-export function resolveTriggerWindow(el: HTMLElement): Window {
-  return el.ownerDocument.defaultView ?? window;
-}
-
-/**
- * Resolve the `Document` that owns `el`, for the outside-click listener and
- * the popup's portal fallback target. A panel popped out into its own OS /
- * PiP window (#1208) renders `SearchableSelect`'s trigger in *that* window's
- * document, not the main tab's — attaching the outside-click listener to the
- * global `document` there means a click inside the child window never
- * reaches it, so the popup never closes; and falling back to the global
- * `document.body` for the portal mounts the popup in the main tab instead of
- * the child window the trigger actually lives in.
- *
- * Unlike `defaultView` above, `ownerDocument` is never null for a connected
- * element, so there's no equivalent fallback to reach for here — this is a
- * direct passthrough, kept as its own named helper (mirroring
- * `resolveTriggerWindow`'s shape) purely so every call site resolves the
- * owning document the same way instead of re-deriving it.
- */
-export function resolveTriggerDocument(el: HTMLElement): Document {
-  return el.ownerDocument;
-}
-
-/**
- * Pure positioning math for `SearchableSelect`'s portaled popup (#1924): given
- * the trigger's viewport rect and the viewport height, decide whether the
- * popup opens down (default) or flips up, and return the `position: fixed`
- * coordinates for either case. Extracted so the flip decision is unit
- * testable without a DOM (`LensPanel.searchable-select-anchor.test.ts`).
- * Takes `viewportHeight` as a parameter rather than reading `window` itself
- * so callers can pass the *trigger's* window (`resolveTriggerWindow`) —
- * necessary for popped-out panel windows, see above.
- */
-export function computeSearchableSelectAnchor(
-  triggerRect: { left: number; width: number; top: number; bottom: number },
-  viewportHeight: number,
-  popupMaxHeight: number = SEARCHABLE_SELECT_POPUP_MAX_HEIGHT,
-): SearchableSelectAnchor {
-  const spaceBelow = viewportHeight - triggerRect.bottom;
-  const spaceAbove = triggerRect.top;
-  // Flip up only when there isn't enough room below AND there's more room
-  // above — otherwise keep the (already clipped-by-viewport, at least
-  // visible-on-screen) default so we never flip into an even smaller gap.
-  const openUp = spaceBelow < popupMaxHeight && spaceAbove > spaceBelow;
-  return {
-    left: triggerRect.left,
-    width: triggerRect.width,
-    top: triggerRect.bottom,
-    bottom: viewportHeight - triggerRect.top,
-    openUp,
-  };
-}
-
-/** Exported for `LensPanel.searchable-select.test.tsx` — rendering it directly
- *  avoids mounting the whole panel (store, discovered data, etc). */
-export function SearchableSelect({
-  value,
-  options,
-  onChange,
-  placeholder,
-  className,
-  displayFn,
-}: {
-  value: string;
-  options: readonly string[];
-  onChange: (value: string) => void;
-  placeholder?: string;
-  className?: string;
-  displayFn?: (v: string) => string;
-}) {
-  const [open, setOpen] = useState(false);
-  const [filter, setFilter] = useState('');
-  const [anchor, setAnchor] = useState<SearchableSelectAnchor | null>(null);
-  const containerRef = useRef<HTMLDivElement>(null);
-  const triggerRef = useRef<HTMLButtonElement>(null);
-  const popupRef = useRef<HTMLDivElement>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const portalContainer = usePortalContainer();
-
-  const filtered = useMemo(() => {
-    if (!filter) return options;
-    const q = filter.toLowerCase();
-    return options.filter(o => o.toLowerCase().includes(q));
-  }, [options, filter]);
-
-  const reposition = useCallback(() => {
-    const el = triggerRef.current;
-    if (!el) return;
-    const win = resolveTriggerWindow(el);
-    setAnchor(computeSearchableSelectAnchor(el.getBoundingClientRect(), win.innerHeight));
-  }, []);
-
-  // The popup is portaled straight to <body> (or the popped-out panel
-  // window's body, #1208) so it clears every `overflow-hidden`/`overflow-auto`
-  // clipping ancestor between the trigger and the viewport — the panel's
-  // scroll container, the floating-panel chrome, the docked-panel host (#1924).
-  // Track the trigger's position while open (capture = also catch ancestor
-  // scrolls, e.g. the panel's own scroll container) and flip up near the
-  // bottom of the viewport instead of overflowing off-screen. Both the
-  // viewport height above and these listeners must use the trigger's OWN
-  // window (`resolveTriggerWindow`), not the global — in a popped-out panel
-  // window the global `window` is still the main tab's, which has the wrong
-  // viewport height and never fires scroll/resize for the child window.
-  useLayoutEffect(() => {
-    if (!open) return;
-    const el = triggerRef.current;
-    if (!el) return;
-    const win = resolveTriggerWindow(el);
-    reposition();
-    const onScroll = () => reposition();
-    win.addEventListener('scroll', onScroll, true);
-    win.addEventListener('resize', onScroll);
-    return () => {
-      win.removeEventListener('scroll', onScroll, true);
-      win.removeEventListener('resize', onScroll);
-    };
-  }, [open, reposition]);
-
-  // Close on outside click — the popup is portaled out of `containerRef`, so
-  // it needs its own containment check.
-  useEffect(() => {
-    if (!open) return;
-    const el = triggerRef.current;
-    if (!el) return;
-    const doc = resolveTriggerDocument(el);
-    const handler = (e: MouseEvent) => {
-      const target = e.target as Node;
-      if (containerRef.current?.contains(target)) return;
-      if (popupRef.current?.contains(target)) return;
-      setOpen(false);
-      setFilter('');
-    };
-    doc.addEventListener('mousedown', handler);
-    return () => doc.removeEventListener('mousedown', handler);
-  }, [open]);
-
-  const display = displayFn ?? ((v: string) => v);
-  const showPopup = open && anchor !== null;
-
-  return (
-    <div ref={containerRef} className={cn('relative', className)}>
-      <button
-        ref={triggerRef}
-        type="button"
-        onClick={() => {
-          setOpen(!open);
-          if (!open) setTimeout(() => inputRef.current?.focus(), 0);
-        }}
-        className={cn(
-          'w-full flex items-center justify-between gap-1 text-left',
-          'text-xs px-1.5 py-1 bg-white dark:bg-zinc-800 border border-zinc-300 dark:border-zinc-600 text-zinc-900 dark:text-zinc-100 rounded-sm',
-          !value && 'text-zinc-400 dark:text-zinc-500',
-        )}
-      >
-        <span className="truncate">{value ? display(value) : (placeholder ?? 'Select...')}</span>
-        <ChevronDown className="h-3 w-3 flex-shrink-0 opacity-50" />
-      </button>
-      {showPopup && triggerRef.current && createPortal(
-        <div
-          ref={popupRef}
-          data-testid="searchable-select-popup"
-          style={{
-            position: 'fixed',
-            left: anchor.left,
-            width: anchor.width,
-            ...(anchor.openUp
-              ? { bottom: anchor.bottom + 2 }
-              : { top: anchor.top + 2 }),
-          }}
-          // Stop pointerdown from bubbling past the popup (e.g. into a Radix
-          // dismissable layer) so picking a value doesn't also close a host dialog.
-          onPointerDown={(e) => e.stopPropagation()}
-          className="z-[120] bg-white dark:bg-zinc-800 border border-zinc-300 dark:border-zinc-600 rounded-sm shadow-lg max-h-[200px] flex flex-col"
-        >
-          {options.length > 8 && (
-            <div className="flex items-center gap-1 px-1.5 py-1 border-b border-zinc-200 dark:border-zinc-700">
-              <Search className="h-3 w-3 text-zinc-400 flex-shrink-0" />
-              <input
-                ref={inputRef}
-                type="text"
-                value={filter}
-                onChange={(e) => setFilter(e.target.value)}
-                placeholder="Search..."
-                className="flex-1 text-xs bg-transparent border-0 outline-none text-zinc-900 dark:text-zinc-100 placeholder:text-zinc-400"
-              />
-            </div>
-          )}
-          <div className="overflow-y-auto flex-1">
-            {filtered.length === 0 && (
-              <div className="px-2 py-1.5 text-xs text-zinc-400">No matches</div>
-            )}
-            {filtered.map(opt => (
-              <button
-                key={opt}
-                type="button"
-                className={cn(
-                  'w-full text-left px-2 py-1 text-xs hover:bg-zinc-100 dark:hover:bg-zinc-700 truncate',
-                  opt === value && 'bg-primary/10 text-primary font-medium',
-                )}
-                onClick={() => {
-                  onChange(opt);
-                  setOpen(false);
-                  setFilter('');
-                }}
-              >
-                {display(opt)}
-              </button>
-            ))}
-          </div>
-        </div>,
-        portalContainer ?? resolveTriggerDocument(triggerRef.current).body,
-      )}
-    </div>
-  );
 }
 
 // ─── Rule display (read-only, clickable for isolation) ──────────────────────
